refactored form post route directory
