+++ src/components/multistepform/MultiStepForm.js
@@ -44,12 +44,15 @@ export default function MultiStepForm() {
 
     // send info to backend on submit
     const handleSubmit = async () => {
-        await fetch("/api/form", {
+        const response = await fetch("/api/form", {
             method: "POST",
             headers: { "Content-Type": "application/json" },
             body: JSON.stringify(formData),
         });
         alert("Form submitted!");
+
+        console.log("response status: ", response.status);
+        console.log("response data: ", await response.json());
     };
 
     return (
